added pheromone trails and decay

+++ Ported/PortedAnts/PortedAnts/Assets/Scripts/Components/PheromonesAuthoring.cs
@@ -21,6 +21,6 @@ namespace Components
     
     public struct Pheromone : IBufferElementData
     {
-        public byte Value;
+        public short Value;
     }
 }
